+++ Library/OcConsoleLib/OcConsoleLibInternal.h
@@ -21,6 +21,9 @@
 #include <Protocol/SimpleTextOut.h>
 #include <Protocol/UgaDraw.h>
 
+#define DEFAULT_COLOUR_DEPTH 32
+#define DEFAULT_REFRESH_RATE 60
+
 typedef struct {
   EFI_GRAPHICS_OUTPUT_PROTOCOL  *GraphicsOutput;
   EFI_UGA_DRAW_PROTOCOL         Uga;